Constants!
Added a paths.inc.php to make our paths easier to deal with. Especially
as we start to move things around in order to conform with the MVC
structure, this makes our links act more like absolute links even
though they are still relative. Check out index.php for an example.
Everything that is in index.php will be moved elsewhere in the MVC, but
for now this is a good example of how to separate out some of the code.